Plugin registration, initialization, and basic functionality.
The WordPress hooks for activation, deactivation, options page rendering,
and actually rendering the GA snippet has been implemented.

The plugin has its own options page under the "Settings"-section. On that
page, there are currently two supported options.

First of, a field for the Google Analytics ID. This is the code you'll get
when you first setup the Analytics account. It'll be displayed on the
Analytics account list, if you missed it the first time.

Second supported option is the domain name. This should be the main domain
you're tracking, including or excluding the www sub-domain. Including the
www sub-domain will prevent Google Analytics from adding cookie data to
resources not located on the same sub-domain. E.g. if you'd use a second
sub-domain as CDN, and want to keep thoses files as light as possible.